started refactoring data handler
